requirements.txt
# Application development
Django==5.2.11
django-anymail[brevo]==14.0
django-htmx==1.27.0
django-import-export==4.4.0
python-docx==1.1.2
qrcode[pil]==8.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1

# Production deployment
dj-database-url==3.1
gunicorn==25.0
packaging==26.0
whitenoise==6.11.0
